Silence all remaining build warnings across gcc, clang and MinGW-w64
Full -Wsign-compare/-Wundef/-Wmissing-prototypes sweep to a clean
zero-warning build on all three toolchains:

- boards/mmc5.c: cast the NES 2.0 PRGRamSaveSize (int in CartInfo,
  guarded nonzero at this site) in the battery-size clamp comparison
  against the unsigned WRAM allocation size [-Wsign-compare]
- hdpack/hdpack_load.c: size the trimmed condition-name scratch buffer
  one byte under the destination so the optional '!' prefix provably
  fits [-Wformat-truncation]
- libretro-common/streams/file_stream_transforms.c: include its own
  header with SKIP_STDIO_REDEFINES (the documented mechanism) so the
  fourteen rf* implementations get prototypes [-Wmissing-prototypes]
- libretro-common/vfs/vfs_implementation.c: make
  retro_vfs_file_seek_internal static - nothing outside the TU
  references it in this vendored subset [-Wmissing-prototypes] - and
  test __USE_LARGEFILE with defined() [-Wundef]
- libretro-common/include/retro_environment.h: test _M_IX86_FP with
  defined() before comparing [-Wundef]
- libretro-common/include/compat/posix_string.h: the retro_strdup__/
  retro_strcasecmp__/retro_isblank__ shims are *defined* for all of
  _WIN32 but were only *declared* for _MSC_VER, so MinGW compiled them
  without prototypes; widen the declaration guard to _WIN32 to match
  the implementations, and make the MSVC-2013 isblank version check
  -Wundef-safe (unchanged semantics: undefined _MSC_VER previously
  evaluated to 0 < 1800)
- Makefile.common: exempt the vendored zlib translation units from
  -Wmissing-prototypes via a pattern-specific CFLAGS variable; zlib
  compiles its LFS-conditional *64 combine functions even when neither
  Z_LARGE64 nor Z_WANT64 exposes their prototypes, and upstream zlib
  does not build with that flag

Verified zero warnings on native gcc, clang and x86_64-w64-mingw32
full builds, plus the HD pack runtime probe and 10x load/unload leak
check unchanged.

+++ src/boards/mmc5.c
@@ -824,7 +824,7 @@ static void GenMMC5_Init(CartInfo *info, int wsize, int battery) {
 			 * (clamped to the allocation) so larger homebrew configurations
 			 * such as the 64K PRG-RAM used by Risa Tracker persist fully
 			 * instead of being truncated to the commercial 8K/32K sizes. */
-			info->SaveGameLen[0] = (info->PRGRamSaveSize < (uint32_t)wsize * 1024)
+			info->SaveGameLen[0] = ((uint32_t)info->PRGRamSaveSize < (uint32_t)wsize * 1024)
 			                       ? (uint32_t)info->PRGRamSaveSize
 			                       : (uint32_t)wsize * 1024;
 		else if (wsize <= 16)
